@@ -1,20 +1,12 @@
 package com.adityap.flashy_createflashcards;
 
-import android.content.Context;
 import android.content.Intent;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
-
-import com.adityap.flashy_createflashcards.models.FlashcardModel;
-
-import java.util.List;
 
 public class CreateCardActivity extends AppCompatActivity {
 
