@@ -2,19 +2,15 @@ package com.adityap.flashy_createflashcards;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 
 import com.adityap.flashy_createflashcards.models.DeckModel;
-import com.adityap.flashy_createflashcards.models.FlashcardModel;
-
-import java.util.List;
 
 public class CreateDeckActivity extends AppCompatActivity {
 
