@@ -1,7 +1,7 @@
 package com.adityap.flashy_createflashcards;
 
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
+import androidx.viewpager.widget.ViewPager;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.adityap.flashy_createflashcards.adapters.CustomPagerAdapter;
